Feature subscriptions

+++ app/src/main/java/co/smartreceipts/android/purchases/subscriptions/RemoteSubscription.kt
@@ -3,6 +3,6 @@ package co.smartreceipts.android.purchases.subscriptions
 import co.smartreceipts.android.purchases.model.InAppPurchase
 import java.util.*
 
-data class RemoteSubscription(val id: Int,
+data class RemoteSubscription(val id: Long,
                               val inAppPurchase: InAppPurchase,
                               val expirationDate: Date)
